project2/requirements.txt
python-telegram-bot==20.8 #эта библиотека нужна для создания телеграмм-ботов на питоне
requests #благодаря ей я могу отправлять запрос ИИ
Pillow #нужна для обработки фото
python-dotenv #для чткния токена из телеги
